SmartResumeMatcher/SmartResumeMatcher/requirements-api.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<4.0.0
python-multipart==0.0.6
pydantic[email]==2.5.3
pydantic-settings==2.1.0
PyPDF2==3.0.1
scikit-learn==1.4.0
numpy==1.26.3
nltk==3.8.1
pytest==8.0.0
pytest-cov==4.1.0
httpx==0.26.0
python-json-logger==2.0.7
psycopg2-binary==2.9.9
python-dotenv==1.0.1
groq==0.4.2
